feat(codemod): add unit tests for codemod cli
