+++ packages/codemod/package.json
@@ -2,9 +2,10 @@
   "name": "@tonic-ui/codemod",
   "version": "2.0.0",
   "description": "Codemod scripts for Tonic UI",
-  "bin": "./codemod.js",
+  "bin": "./codemod-cli.js",
   "files": [
-    "codemod.js",
+    "codemod-cli.js",
+    "codemod-main.js",
     "dist"
   ],
   "publishConfig": {
